Teach Identity.hasSession to fall back to spid
Version 2.x of this SDK has a check that if a return value from
hasSession contains an “error” key with child “type”: “LoginException”,
it will repeat the request, but send it directly to SPiD.

There was a bug in this version. A call to fetch only errors when there
is a network error, so the error handling to check for “LoginException”
was placed in the wrong spot. I have moved it, and also created a unit
test to simulate this and verify that such a response from hasSession
should in fact retry the request to SPiD.

Thanks to Terje Andersen for reporting the bug.