identity: Send correct autologin parameter
The hasSession and SPiD backends expect the autologin parameter to be
either ‘0’ or ‘1’ (and not false/true). This used to be how the Identity
class worked, but a recent refactoring introduced a bug there.

This commit refactors the unit tests for hasSession, adds tests for
autologin, and fixes the issue.