SDKError: change unit test for istanbul/babel
When jest runs with --coverage, it uses istanbul for that, and there are
multiple issues reported between jest/istanbul/babel. I’m not entirely
sure what happens and which of these libs need a fix, but the net effect
is that we can’t inherit core classes like Error or Array when coverage
is enabled. Added a TODO to flip this back when (if) the issue is fixed
upstream.