identity: Make hassession unit tests timezone-independent
One test assumed a CET timezone, and failed on Travis. Let’s refactor
that error message, so that we just fetch a type and value (preferably
using .valueOf()), and use that instead of .toString(). This does
introduce a bit more logic that I would have wanted here, but.. I can
live with it.