identity: Add unit tests for _maybeSetVarninshCookie
This made me find an issue, where we sometimes need to reset the cache
used by the Identity instances. The issue I found was when
enableVarnishCookie() is called; if hasSession() is called *before* this
function is called, it’s important to not use the cached version, so we
refactor this logic into an _initCache function that we call both from
the constructor (to keep the existing functionality) and now also from
enableVarnishCookie().

Another thing I disliked a bit in the existing code was that in
_maybeSetVarnishCookie, we modified the input data. This just felt
smelly, so I changed from re-writing sessionData.baseDomain to having an
internal const instead.