identity: refactor getVisitorId and add unit tests
getVisitorId is now marked async and is simplified. I think it was
trying before to extract a visitor uid also in the case where hassession
returns an error, but this would not have worked in its previous
implementation. It doesn’t in this one either, but at least we’re not
fooling ourselves anymore ;)