Reverse login in urlMapper helper function
This is really a workaround for the case where a client needs the
url-polyfill feature in core-js for browsers like IE11 where ‘new URL’
doesn’t work according to spec. Specifically, if called as

new URL(‘foo’)

it will be perfectly happy and return a url using window.origin as the
base url. Even more specifically, since we used to call new URL(‘PRE’)
and expect it to fail, this was a problem. But reversing the logic is so
simple that I’ll choose that here. Basically, if called with PRE like we
do, we’ll check if there’s an entry in our env map *first*, so we’ll
never hit the bad ctor call at all.