Ensure we clear cache when logging out
Fixes #68

Bah! The Identity.logout() and Identity.enableVarnishCookie() functions
both called _initCache which doesn’t clear the existing key from
localStorage. This caused these function to *think* the cache was
cleared, when in fact it wasn’t. Weirdness ensues… Issue is fixed and a
test was added to make sure it doesn’t happen again.