identity: Fall back to spid-hassession for hasSession()
For migration after session-service is put into production, let’s fall
back to polling spid-hassession for all the times when users don’t have
a valid cookie in the id.whatever domain.

This patch checks explicitly for the 404 return code from the
session-service /session endpoint, and in that case falls through to the
good ‘ol spid-hassession-and-then-spid sequence for checking whether a
user is logged in.