cache: Handle when just accessing webStorage throws
On for instance Safari 10 in Private mode, the simple act of accessing
window.localStorage will throw an exception. This commit adds a test for
this case, as well as change the constructor of the (internal) Cache
class to take a provider function instead of an actual storage object.
This is just to keep the complexity of dealing with that possible
exception inside the Cache class (in the ‘webStorageWorks’ helper
function).