hasSession: Rewrite to use a “singleton Promise”
This commit proposes to create a Promise instance whenever hasSession is
first invoked, and then, if any successive calls are made to hasSession
before the first one is done, we simply return a reference to that same
Promise instance. This both prevents unnecessary network calls, and also
reduces the chance of race’y situations both with and without the ITP
dialog.